% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rproject-package.R
\docType{package}
\name{Rproject-package}
\alias{Rproject}
\alias{Rproject-package}
\title{Rproject: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\details{
Rproject: A package to analyze 10x multi-omic data.
This package provides a set of functions to perform a complete
analysis workflow as described in the course project.
}
\author{
\strong{Maintainer}: Jane Doe \email{giorgia.scelza@edu.unito.it}

}
\keyword{internal}
